Remove calls to python2 support/libs.
Checking for python 2 support in the python 3 branch seems unnecessary, so I
removed that.  The StringIO import in the lint function setup was changed to
import io, the `except` syntax was fixed, the key function parameter to sort was
changed to pass by keyword and changed the behavior of the lambda to function
properly there.  Pylint, to my knowledge, does not have python 2 support so I
removed that particular code entirely.  Also, a call to the python 2 file
builtin was replaced by a call to open.

Added a set of g:pymode_syntax to stop an error I did not fully track down.